+++ cli/commands/check/subcommands/remove_subdue.go
@@ -35,7 +35,7 @@ func RemoveSubdueCommand(cli *cli.SensuCli) *cobra.Command {
 				return err
 			}
 
-			fmt.Fprintln(cmd.OutOrStdout(), "OK")
+			fmt.Fprintln(cmd.OutOrStdout(), "Removed")
 			return nil
 		},
 	}