+++ cli/commands/check/subcommands/set_subdue.go
@@ -62,7 +62,7 @@ func SetSubdueCommand(cli *cli.SensuCli) *cobra.Command {
 				return err
 			}
 
-			fmt.Fprintln(cmd.OutOrStdout(), "OK")
+			fmt.Fprintln(cmd.OutOrStdout(), "Updated")
 			return nil
 		},
 	}