+++ cli/commands/event/resolve.go
@@ -33,7 +33,7 @@ func ResolveCommand(cli *cli.SensuCli) *cobra.Command {
 				return err
 			}
 
-			fmt.Fprintln(cmd.OutOrStdout(), "OK")
+			fmt.Fprintln(cmd.OutOrStdout(), "Resolved")
 			return nil
 		},
 	}