+++ cli/commands/handler/delete_test.go
@@ -46,7 +46,7 @@ func TestDeleteCommandRunEClosureWithFlags(t *testing.T) {
 	require.NoError(t, cmd.Flags().Set("skip-confirm", "t"))
 	out, err := test.RunCmd(cmd, []string{"test-handler"})
 
-	assert.Regexp("OK", out)
+	assert.Regexp("Deleted", out)
 	assert.Nil(err)
 }