+++ cli/commands/hook/update.go
@@ -52,7 +52,7 @@ func UpdateCommand(cli *cli.SensuCli) *cobra.Command {
 				return err
 			}
 
-			_, err = fmt.Fprintln(cmd.OutOrStdout(), "OK")
+			_, err = fmt.Fprintln(cmd.OutOrStdout(), "Updated")
 			return err
 		},
 	}