+++ cli/commands/silenced/update.go
@@ -47,7 +47,7 @@ func UpdateCommand(cli *cli.SensuCli) *cobra.Command {
 				return err
 			}
 
-			fmt.Fprintln(cmd.OutOrStdout(), "OK")
+			fmt.Fprintln(cmd.OutOrStdout(), "Updated")
 			return nil
 		},
 	}