+++ CHANGELOG.md
@@ -23,6 +23,7 @@ by adjusting the max_output_size and discard_output properties.
 - Removed unused workflow `rel_build_and_test` in CircleCI config.
 - Moved the `Provider` interface to `api/core/v2` package.
 - Moved the `Authenticator` interface to `backend/authentication` package.
+- Updated confirmation messages for sensuctl commands: `Created`, `Deleted` and `Updated` instead of `OK`.
 - Exported some functions and methods in the CLI client.
 
 ### Fixed
@@ -49,7 +50,7 @@ authentication provider.
 - Modified private generic store methods as public functions.
 - Improved logging for errors in proxy check requests.
 - Updated Go version from 1.10 to 1.11.4.
-- Changed keepalive event to include check.outputé
+- Changed keepalive event to include check.output
 
 ### Fixed
 - Fixed a bug where `sensuctl edit` was not removing the temp file it created.